[IMP] sale_automatic_workflow_payment_mode: improve test coverage
